fnc_arcor_orders_subscriptor/requirements.txt
functions-framework==3.*
google-cloud>=0.34.0
google-cloud-storage>=2.7.0
python-dateutil